refactor: move card to its own component
